Added the http and util module!
The http module currently contains a server.py which has classes to
create a MultiHandlerSingleThreadHTTPServer along with a mutable
MultiHandlerDispatcher for HTTP requests. Added tests for that.

The util module contains a type.py that houses the Singleton class type.
This is used by the http module Global Handler Registry to let the
Dispatcher retrieve the correct handler. Added tests for that too.